Put Time#floor and Hash#transform_keys monkey patches in spec-only refinement. To better expose version issues.
